+++ executioncontext/api/src/main/java/org/jetbrains/bsp/bazel/executioncontext/api/ExecutionContext.java
@@ -1,0 +1,7 @@
+package org.jetbrains.bsp.bazel.executioncontext.api;
+
+/**
+ * <code>ExecutionContext</code> base class - you need to extend it if you want to create your
+ * implementation of <code>ExecutionContext</code>.
+ */
+public abstract class ExecutionContext {}