+++ executioncontext/api/src/main/java/org/jetbrains/bsp/bazel/executioncontext/api/entries/ExecutionContextEntity.java
@@ -1,0 +1,11 @@
+package org.jetbrains.bsp.bazel.executioncontext.api.entries;
+
+/**
+ * Base <code>ExecutionContext</code> entity class - you need to extend it or <code>
+ * ExecutionContextListEntity</code> or <code>ExecutionContextSingletonEntity</code> if you want to
+ * create your entity.
+ *
+ * @see org.jetbrains.bsp.bazel.executioncontext.api.entries.ExecutionContextListEntity
+ * @see org.jetbrains.bsp.bazel.executioncontext.api.entries.ExecutionContextSingletonEntity
+ */
+public abstract class ExecutionContextEntity {}