+++ executioncontext/projectview/src/main/java/org/jetbrains/bsp/bazel/projectview/model/sections/ProjectViewSection.java
@@ -1,0 +1,14 @@
+package org.jetbrains.bsp.bazel.projectview.model.sections;
+
+public abstract class ProjectViewSection {
+
+  protected final String sectionName;
+
+  protected ProjectViewSection(String sectionName) {
+    this.sectionName = sectionName;
+  }
+
+  public String getSectionName() {
+    return sectionName;
+  }
+}